requirements.txt
fastapi
uvicorn[standart]
colorama
sqlalchemy
psycopg2
python-dotenv
requests
pytest-factoryboy
pytest-cov
python-multipart
alembic